Misc fixes
Supersonic Boots description said it takes 10000 souls to upgrade, really takes 20000
All crests dropped as a stack of 2, but had a max stack size of 1 so they couldn't be split or moved seperately.
Enemy Lightning Bolt Ball 3 and 4 could be controlled by the player if they were holding use on a (magic?) weapon that is constant use, like Last Prism, Rainbow Rod, or Divine Boom Cannon.