Artorias Event and Consumable Soul Bug
Removed an example custom action in the Artorias event that was simply to showcase various things custom actions could do, and was not intended for actual gameplay.

Added some debug code and a potential fix for the issue where consumable souls drop multiple times in multiplayer with 2+ players, but it needs further testing to ensure it actually resolves the issue.