Reversing gravity no longer flips your screen
Added an experimental setting to disable the screen flip when gravity is reversed. Still experimental, so it defaults to off for now. A number of visuals (NPC house icons, sword glowmasks, etc) need to be edited to not flip.

Added a setting to disable inverted aim when gravity is flipped. This feature is very unfinished! Hover Text, Grapples and Laser weapons all still act like your cursor is on the opposite side of the screen. Maybe your aim *should* just always be inverted when flipped as a downside though?