NewText changes, in prep for porting
UsefulFunctions.ServerText is now UsefulFunctions.BroadcastText to clarify it's not just for server usage but for any time you want to display a line of text to all players.

Changed most NewText calls to use BroadcastText, since NewText can cause crashes in 1.4 if used in a server context. BroadcastText is safe to use anywhere except in functions that run only on one client like in item "use" functions. Use NewText there instead.

This was done manually, and was not done for items or other things that are client-side only and are only supposed to display a message to one client (NewText is intended to be used there).