Tweaks and fxes
Yes, fxes. It's only part of a fix, so it doesn't deserve to have that typo corrected.

Updated Humanity description to be more clear about its purpose to non-BotC players
Slimy Saddle now is a guaranteed king slime drop, since there is now an intended way to skip the jump gate by spawn anyway

Blocked Rod of Discord and mech spawners, but not picksaw and slime hook because TML's new drop system is final proof that we are in fact living in hell. Whatever, they can only drop in normal mode for now anyway.