Fixed Shaman Elder Dialogue
Color tags don't seem to work in 1.4, and text formatting changed a bit. Also split some of the Shaman Elder's text into an extra page, because his dialogue didn't fit in the box anymore. Old version is commented for future reference just in-case we find a way to make color tags work again.