More fixes + crates
Added Soul Shekel crates, which take the place of removed gear for players. When removing an item in the future, add it to the LegacyName list above it.
Burning Stone now works with Reflection Shift
Reduced damage of Fire Lurker's lost souls
Added Meteor Shot crates
Buffed the Dark Trident significantly to be on-tier with similar options (its slow charge time balances its high damage, and its spear mode only inherits a fraction of it)