Chloranthy Ring buff + new SHM upgrade
Alright, I finally gave this idea a swing. Tested and tweaked the stats quite a bit to find something that felt good and looked noticeable for each ring compared to the base roll while aiming to keep the benefits conservative (boosting the iframes stat just a few higher (22 and 26 for example) it was easy to see how it could get OP fast, so I brought both of them down from my initial tests)

-Chloranthy Ring I now enhances the dodge roll (momentum is retained slightly better after dodge, dodge cooldown reduced and +3 iframes)
-Added Chloranthy Ring II: boosts dodge roll effects slightly more (+momentum retention, no dodge cooldown, and +2 iframes)

To help balance these boosts, I brought down the stamina recovery speed of I, from 20% to 15%, and ring II gives 25%, plus gives a 10% chance to evade hits on contact (black belt ability).

Ring II is obtainable after defeating Attraidies.
The rings cannot stack.

If there is anything I would add to these, it would be to buff the distance of the mid-air dodge with each ring slightly, as the rings currently don't offer any increased mobility in the air, but I couldn't make heads or tails out of how that was working. :D

As always, I'd consider this a 1st pass on the idea, so looking forward to discussing, gathering feedback and iterating further as need be.

Zeo or RC, if you're up for boosting the mid-air dodge a bit for these, that'd be sweet!