Downloader Rework
Redid a lot of it. It now runs extensive checks to see if the downloaded copies of the map or music mod were corrupt or incomplete. For the world it follows the same process vanilla does to determine if it is corrupt: It prompts the game to load the worlds metadata, and verifies its hash. For the music mod it simply checks if the hash inside the mod file matches the hash *of* the rest of the file. Both of these methods are far more robust than simple file size checks.

If it finds a corrupt or incomplete file, it discards them and tries to redownload a fresh one. In the case of the music mod, the download text on the main menu will say "Music Mod download failed, click to retry".

If the world download fails it re-tries, but only up to 3 times. If it fails all 3 times it will display a popup informing users that they need to download it manually from our discord instead, with a link displayed.

It also runs a check on startup to see if the *current* copy of the base map is missing or corrupt, and if flags a new download as needed. Now users who previously got a corrupted version will now have it auto-detected and fixed.

Finally, added map download progress indicator to the main menu. This shows the progress of the update and alerts users that a new map update has happened once it finishes.