Event Code Updating
Renamed several variables in the events class:
DisabledEvents>EnabledEvents, since its name was very counterintuitive. Events in this list are being checked every tick to see if the player is in range.
ActiveEvents>RunningEvents, since the events in EnabledEvents are "active" too, in that they exist in the game world. This makes it more clear that these are the events which have been tripped and whose code is currently being run.

Replaced several individual checks for bosses with tsorcRevampWorld.BossAlive to simplify code

Commented out deprecated recipe code