Finished crafting simplification
This includes a pass on: Ranged, Throwing, Summon

Another goal/strategy I forgot to mention is putting a focus on the soul currency as the primary crafting mechanic, so grinding for other random ingredients is less of a grind (once you can obtain 1, you can obtain 30 it just takes a long time).

In some cases I left high numbers as they were when it seemed sensible for the given weapon.

In some cases where I lowered the other ingredients a lot, I bumped the soul cost a bit to compensate - in general, just wanted to make the crafting and all the material juggling simpler and less of a pain.

The last goal I had was discoverability. Ensuring mod stuff will hopefully get noticed easier based on the starting ingredient.

Also with less overlap between common ingredients, the list of items when examining something like cobalt bars or souls of night should be smaller and easier to digest.

- Several mod enemies now have a chance to drop souls of night, light and flight