Additions and Changes
Added visions of the bosses required to kill Gwyn to the walk down in his tomb

Added Radiant Glimmer, a new endgame mage weapon. VFX are not representative of its final form, i'm just experimenting a bit right now.

Enabled caching on the shader for marilith's firewalls for a slight performance boost

Marilith's firewalls now simply subtract health instead of "hitting" the player to prevent rapidly activating on-hit effects like the hurt sfx and star cloak

Added Artorias and Witchking to Boss Checklist

Added GetPlayerHandOffset, which lets you get the exact position of the players hand in the world (for things like drawing effects right on it)