New Laser Visuals and Cursed Flamelash Rework
Laser visual rework is complete! This heavily changes the visuals of pretty much every laser in the game. Notable examples are the Destroyer and Attraidies fights, as well as the Crystal Spectrum and Arcane Lightrifle weapons

Also, Cursed Flamelash is now the Cursed Tormentor, and it functions in a more unique way