Experimental Hunter change, considering applying to Sorrow and the Rage
The Hunter now takes 500 burst damage when enraged, but damage dealt during this phase is also increased so that there's a risk/reward element to enraging the beast or not, instead of it being solely a bad thing.

I think this may end up being a lot more engaging and rewarding for folks, allowing the choice of a safer battle (avoiding the enrage) or a more challenging but potentially more satisfying and quicker one.