Little Gwyn buff
Too many reports of too easy lately.
The main reversion of Marf's nerf is the raining black skull attack damage. This was supposed to be high to make running away very dangerous and encourage the player to stay close (where his other attacks are more of a threat) or use the dome for shelter. Since it can be avoided it seems alright. This attack should be scary basically, more than most.

To go further with that, I added 2 damage buffs when the sword is destroyed and added some broadcast text.

I also raised the damage of attacks that are strongly telegraphed like the green and red flame thrower attacks.