Slight worm nerf
these enemies are just kinda annoying amiright? at least if they're less of a pain they'll be alright. but the whole waiting for them to pop up forever and getting one hit on them then waiting again is a gameplay loop I think we should minimize in favor of more interesting enemy encounters.