Increased spawn rate and max spawns by 10% for each player health threshold
This should help with the reported low spawn rate and low enemy count of the jungle, dungeon and fire temple where our spawn rate adjustments based on health could have been a factor. Easy to adjust though if that tips the balance too far in the other direction.

Also made ringed knights only spawn in the underground to increase uniqueness of the enemy/biome