Sound and volume tweaks for Lethal Tempo SFX
I used Audacity to trim the SFX to be quieter and not so long so they don't overlap on top of each other when hitting enemies multiple times. Goal is for the sounds to blend with the main weapon swoosh sound and to be short to go along with the quick nature of a "clang" sound hitting an enemy.

Let's test these out and make further changes as needed. I believe they are a conservative improvement.