Update tsorcGlobalProjectile.cs
Part of a combination of changes. Adding a "slow" debuff to Baby Slime which should slow the movement speed of mobs (added as a utility power of Baby Slime Minion).

Adding a "Frostburn" Debuff to Flinx minion, 1 second duration which should equate to 8 points of damage per second. Will change direct  contact damage down to 1 point to compensate. Allows Flinx minion to be an early mode Frostburn debuffing option as a summoner to deal with early mode high defense creatures.