Update AshenLord.cs
The float scale should adjust the entire dragon (segments included) visually to be smaller. While the segments in the code appear to be at "set" locations when creating (as each one as animation on it's own) the float scale should be the device/code (as i understand) that then scales the entirety of the formed dragon as a whole piece. It's functions would remain the same.  

In summary: The float scale should adjust the total dragon, positional segments (arm/legs) included (x and y locations), to be smaller.