Update SunsetQuasarMinion.cs
This should increase/double the speed and height of the Sunset Quasar Minion.  Being a ground minion (As mentioned in my discord threads) it suffers from being able to get to and stay on flying targets. This should help to address this issue and make the furry critter more of a threat without having to adjust and damage parameters. 

The changes should only affect it's movement when targeting and chasing enemy (so essentially increased it existing combat movement parameters)