Q.O.L. Keep Station Buffs on Respawn
Because the various buff stations are back in the main town, after death you have to teleport back to reapply the free buffs. This can get annoying in a mod where death is so common. With this change as long as a player has used a station once before after entering the world, they will not lose it after death. Honestly should have been vanilla behaviour imo.